hide std_to_otel function to convert python logging severity to otel severity
